requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.0
flask>=2.3.2
pydantic>=1.10.7

# Database
pymongo>=4.3.3
redis>=4.5.4
mysql-connector-python>=8.0.32

# Data Processing
numpy>=1.24.2
pandas>=2.0.0
scipy>=1.10.1
statsmodels>=0.13.5
matplotlib>=3.7.1
seaborn>=0.13.0

# Trading & Market Data
rqdatac>=2.9.0
tqsdk>=3.2.10

# Task Scheduling
apscheduler>=3.10.1

# Utilities
loguru>=0.6.0
PyYAML>=6.0
tqdm>=4.65.0
setuptools>=67.6.1

# Development & Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
IPython>=8.12.0 
openai>=1.73.0